fix: Score type in OthersRatingComment.

+++ backend2/src/main/java/com/cmoney/backend2/forumocean/service/api/rating/OthersRatingComment.kt
@@ -17,7 +17,7 @@ data class OthersRatingComment(
     @SerializedName("memberId")
     val memberId: Long?,
     @SerializedName("score")
-    val score: Double?,
+    val score: Int?,
     @SerializedName("reviewTime")
     val reviewTime: Long?
 )
